Init Eigen testing
In the eigen_test project I mainly try to find a suitable method to allow for
a (linear algebra) library agnostic approach, such that libraries can easily
be swapped without altering the main code.
The inputPath in the Setup struct has been changed to comply with the new
location of the input files.