added functions to parse ranges
By calling getFirstValue the user will obtain the first value in a range if a range is specified. If a direct value is specified the function will simply return that value. This function is used in the constructor of the WorkingConditions class.
Note that complete range parsing as necessary for the job system still has to be implemented (but it is very similar in terms of regular expressions).