switched from raw pointer to unique_ptr
Smart pointers will be used to ensure exceptions will be handled without leaking any memory.
Another option would be to add a try catch block to constructors that might be prone to leaking memory when an exception is thrown. The deletion of memory can then be properly handled in the catch block.