added Notify function to Log class
The user can now also use the log class to display notification instead of just warnings and error.
SmartGridSetup values are initialized to zero.
The updatedGasTemperature event in the WorkingConditions class has been split into two events (as in the Matlab implementation).
Runtime outcome is now printed to std::cerr (for consistency).