added gas mixture implementation with trait class and without indexing
This implementation of GasMixture is similar to 'trait_test' in that it uses trait classes to specify the Boltzmann and Chemistry types. However, instead of states indexing their children, this implementation simply adds child states to a vector (closer to the Matlab implementation). It differs from the Matlab implementation in that it immediately creates missing parent states