major commenting + bug fixes + cleanup
Ionic states are now added to their own tree within Gas.h.
Removed 'ionic' as a StateType.
Switched from a vector of Gas objects to a vector of pointers to Gas objects to fix a bug in the GasMixture class that occurred when multiple gasses were added (due to copying of the vector on resize, which invalidated pointers).
Intermediate input structures have been moved to the InputStructures header.
Major commenting of the templated base class structure (Gas, State, Collision and GasMixture).