implemented state property loading from direct values and functions
Main feature is the implementation of state property functions.
More property functions need to be added, for now it's just the functions required for the default Nitrogen setup.
A new LXCat file containing a lot of vibrational collisions for O2 has been added.
Pointer to workingConditions is now passed to ElectronKinetics constructor in order to access its arguments while assigning state properties.