bugfixes and commenting
Fixed a bug in the Parse::removeComments function where it would not work on strings that do not have and endline character (e.g. the last line in a file or a line obtained through std::getline).
Fixed that states in a property file can also have wildcard characters.
Commenting of the newly implemented loading of state properties.
StateEntry objects can now be printed to standard output.