Implementation of linear operators and matrices and matrix inversion
Added Elastic, Field, CAR and inelastic operators and matrices to the ElectronKinetics class.
Using an LU decomposition algorithm with partial pivoting to solve the linear system of equations.
Added loading of the CAR gasses into the EedfGasMixture class and the checking of their requirements.
Implemented deduction of the total elastic and total cross sections.
Added the computation of superelastic cross sections.
Added the possibility to print two vectors such that output can be piped to gnuplot for plotting.