implemented Hessenberg reduction
The reduction combined with the already present Hessenberg solver allows for fast inversion in the case that the matrix is already near Hessenberg.
Two extra variations of the reduction algorithm have been implemented.
The first (dubbed optimal) is more efficient in the case of a sparse superelastic band structure (no clustering of bands in lower triangular matrix).
The second introduces partial pivoting to promote low rounding errors in most situations.