cmake usability update
The required libraries to build with Intel MKL / OpenBLAS are now present in the 'dependencies' folder.
CMake can now be configured to build either with Intel MKL, OpenBLAS or pure Eigen, by passing a flag.
To compile:
- cd build
- cmake -D<BACKEND_FLAG>=ON ../
  - Where BACKEND_FLAG=USE_MKL/USE_OPENBLAS
  - The flag can also be omitted to build with pure Eigen.
- cmake --build . --target loki --config Release -- -j <NUM_JOBS>
  - Where NUM_JOBS is the number of jobs to be run simultaneously when compiling -> just choose the number of physical cores.
NOTE that the libraries supplied are for linux; compilation, for now, will probably not work on another operating system.