added WASM build of LoKI-B
LoKI-B can now be compiled to WebAssembly
The build is working when run using node ('cd wasm_build/' 'node node.js')
It does not yet work when run in the browser ('npm install http-server -g' 'cd wasm_build/' 'http-server')
since it cannot access the file system. Thus another method needs to be implemented to supply loki with the
required files (for this particular case to run).