Added "reverse_also" (bool) to "reaction" elements
In LXCat files, the usage of <-> (as opposed to ->) indicates that
also the backward process must be handled. This cannot be inferred
from the JSON files, since these do not have a separator (only "lhs"
and "rhs" objects). For now, fixed the Nitrogen files by adding
"reverse_also": true|false to the reactions. (The reactions e + N2(X,v=0)
-> e + N2(X,v=10) and all rotational processes have true, the others
false).