Implemented loadCollisionsJSON.
LoKI-B can now be started with a JSON input file. See for example
Input/default_lokib_setup.json, and change the LXCatFiles array
to use the json files. The results are the same.

NOTE that some changes are needed to the JSON format that was initially
'v4', and various questions remain:

 * The JSON versions of the LXCat files do not expose the difference
   between '->' and '<->' in processes.
   (I added reverse_also to the Nitrogen files to communicate this for the time being.)
 * How is stoichiometry handled? Is 'e + e + X' translated into
   '2 e + X'? Or should that be the case? How is this in the MATLAB version?
 * In the JSON, "v" is inconsistent. Sometimes it is a number, sometimes a string
   (Example: "10+" is used in Input/Nitrogen/JSON/v4/N2_LXCat.json)