Further 'JSONification of LoKI-B".
When the Simulation object is now constructed with a JSON object,
no Setup object is ever created. Instead the JSON object and its
entries are recursively passed down to the Simulation members.
This also means that most Parse members that were added before
to support JSON can now be removed again. Parse:: members are still
used in a few places that will be handled later (for example: a
range is now constructed from a string, in JSON this could be made
a proper object). The State(Entry) construction is now also done
directly from JSON, but note issue #1, issue #1 and issue #3 that
describe work that must still be done.

When JSON is used, tiny differences can be ibserved (15th decimal,
typically), probably related to the handling of doubles by libstdc++
vs. JSON.

This changeset fixes issue #4.