Make single and multi-runs behave the same.
Also the single-run case (Range-aware values are plain numbers)
are now run using the JobManager mechanism. This makes the code
much more uniform and avoid the need for special-case parsing
like getFirstValue.

To make this possible, the Range class has been updated to behave
well for n=1 (value==start), allowing ranges to be used for
single-value cases. In addition, member prepareFirstJob() has been
added to the JobManager, to do post-construction inialization of
relevant values (E/N or T_e, typically, but more can be added more
easily after this patch.)

This change allowed a quite substantial reduction of the amount
of code in Parse.h and results all in all in a better localization
of concepts.