Exception safety fixes, typename improvements
This changeset implements a number of exception safety fixes,
replacing the usage of vector<Y*> (in combination with delete
statements in destructors) with vector<unique_ptr<T>>. This
ensures that memory will be reclaimed also when object creation
fails when the constructor has not been executed completely.
For readibility purposes, the gas and mixture classes provide
typenames for the relevant types that are obtained from the
traits classes. (As an example, in Gas<Traits> 'State' is now
used instead of 'typename Trait<TraitType>::State'.

Spelling fix: gasses -> gases in a few places.