Make Job::callback a std::function<void(double)>.
The callback is now a standard function that accepts a double and
returns void. When a jon is added, the choice is made to make
this function a member function call on the particular workingConditions
object (using bind), but other choices would be possible as well.

This makes the JobManager and associated classes fully generic,
essentially loki-b independent.