Revert 'reverse_also' patch
Revert the addition that was introduced in changeset 340fad76587437b88ad38b4b65f759a3f6046f1c.
Instead of reverse_also, use field superelastic, which has the
same meaning.

TODO: superelastic is a bad name. It is also possible that the
forward reaction is superelastic, and that the backward process
should (or should not) be taken into account.