Revert 'reverse_also' patch
Revert the addition that was introduced in changeset 52a1c93277bdafea33008f7f93f172e124e06959.
Instead of reverse_also, use field superelastic, which has the
same meaning.

TODO: superelastic is a bad name. It is also possible that the
forward reaction is superelastic, and that the backward process
should (or should not) be taken into account.