Moved EedfState member data to EedfGas.
In the EedfGas, we now have two variables map<StateBase*,vector<EedfCollision*>>,
so the EedfState is essentially just a normal State<Boltzmann> (and then,
a simple GasBase::StateBase). That will be used separately. In addition,
the change extraCollisions -> collisionsExtra was made. This facilitates
'pattern recognition' while analyzing code that uses collisions.