Do not use Traits, but direct template params.
 - instead of relying on traits classes, use the state and gas types
directly in template specializations. This avoid a (near) circular
dependency in the previous code and is (IMHO) more straightforward.
 - introduce state types root and charge. Every gas has a root state
nenber, all states are direct or indirect children of this state.
This avoid the duplication of code in the gas and state classes
(e.g. checkPopulation) and avoids the need for separate neutral
and ion state trees.