Use the new state entry parser.
This catches many more mistakes than the old one. See the results
of test_reaction_format for an overview. In addition, the stoichiometric
coefficients are also parsed for the LHS, and it is chacked that
the values are 1. (Before, smth. like '2 e + N2(X)' would be happily
parsed and interpreted as 'e + N2(X)'.

For now, the new parser is enabled by C++ macro LOKIB_USE_OLD_STATEENTRY_PARSER,
see file EedfGasMixture.cpp. A next patch will unconditionally
use the new parser and get rid of the old.

@DAANBOER, please have a look.