More generic state array creation.
When the new parser is used, all particles that appear in a reaction
are recognized, including the electrons. These are now removed from the
state array just before calling createCollision. Also the LHS
stoichiometric coefficients are noew parserd. These should be used later
to invalidate reactions such as '2e + X -> ...' or 'e + 1X -> ...'.
Later it makes more sense to construct the reactions such that all
participants are recognized and stored in the Collision object. The
EedfCollision can then itself remove the electrons if that is easier.
This has not yet been enabled in this patch because this requires the
creation of an electron gas/state, which is problemetic at present for
LoKI-B (for reasons beyond my understanding).