Added (disabled) support for FPU exceptions.
Added support for using <cfenv> functions to trigger an abort if
divisions by zero, overflows or invalid numbers (NsN's) are
encountered.

See disabled macro //#define LOKIB_ENABLE_FPU_EXCEPTIONS in
source/main.cpp.

This code reveled that ElectronKinetics::evaluateFieldOperator
produces a NaN for the first element of vector g_E, which is later
repaired by setting it to zero.

This facility may be useful, but must first be implemented in
a portable way (use feraiseexception etc. instead of glibc's
feenableexception). And of course, it must stay a non-default
optional.