Take into account the electrons.
Also handle the electron entries in processes, and create an
electron state and an electron gas. To make sure that the eleectron
is modeled the same everywhere, added static StateEntry member
electrinENtry(), which returns the entry that should be used to
create/describe electrons (at present a 'charge' state with charge
id "-" and gas name "e").

The EedfCollision keeps a list of heavy states that appear on both
sides of the equation, just like before. This change results in
more straightforward descriptions of reactions, less tailored
specifically to the need of a Boltzmann solver. Without ever having
seen Loki-C, this should also facilitate integrating LoKI-C
(chemistry) at some point.