Added a small tool for matrix -> xpm.
This accepts a loki::Matrix and creates an XPM file. Negative,
positive and zero-valued matrix elements appear in red, blue and
white. This is useful for quickly understanding the sparsity pattern
of a matrix, and was used to create the picture in issue #7

TODO: xpm is sooooooooooooooo 90's, but hey: so am I.. There may
be another format that allows equally easy and efficient generation
of pictures with (very) small color depth.