Added N2 LXCat files in new format
The new documents are objects with two properties:
states: The list of states involved in the reactions
described in the document. Additionally, all states
are assigned a (string) id.
cross_sections: The list of cross sections. All
states in lhs/rhs of reactions are now referenced
by their newly given ids.