Use StateEntry id's.
StateEntry now has a string member m_id. This is read from the
input file and is essentially the unparsed string, like "N2(X)".
The mixture creates a <id,State*> map. When ensureState is called,
it will look for the State based on the id, and return it when
found. Otherwise, the state is created and added to the map.

loadCollisions (JSON version) now first reads/creates all States
that appear in the "states" array, subsequently it creates the
reactions based on the cross_sections array.

The code now assumes the v5-with-ids JSON version. Two more
changes have been made in thiese input files:
 * "e^-" -> "e" according to folklore. This also improves readability
 * "count" is now optional. When missing, 1 is assumed. This
   removed a lot of cruft from the JSON files.