Added loki_bindings target to CMake
loki_bindings declares a simple WebAssembly binding to run
a LoKI simulation.
The target is built automatically when emscripten is
detected.
Added <sstream> to Parse.h to fix a compiler error using
em++.
For now the simulation is run using the preloading of input
files. This should later be changed to allow users to upload
their own input file. However, for the database files it might
be a proper short-term solution.