Added plotting functionality
The handle_results callback in C++ calls the plot
function in Javascript. The pointers and sizes of
the x and y contiguous arrays are passed as parameters.
The plot function then reads the data directly from the
wasm heap into ArrayFloat64 structures. These structures
are transformed to a Javascript array of {x:..., y:...}
objects, which is plotted using Vega-Lite.