Allow other output mechanisms than file writing.
Output: this is now a base class that provides interfaces for
producing output, but does not assume that files are written to
the local file system. That output type is implemented in derived
class FileOutput.

WIP: class JSONOutput appends output to a JSON object that is
passed by reference when the simulation is set up. This 'root object'
corresponds more or less with the output directory, data for
specific cases are written to subsections (~ subFolder) or appended
to objects that are inside the root object (~ lookup_table). Note
that at present only the setup file and the eedf are written into
the JSON object, for other outputs (rate coefficients, lookup tables)
the operations still need to be implemented --- see Output.cpp.