Pass the 'file exists-handler' to the Output ctor.
This makes it possible to avoid two event objects for the sole
purpose of allowing the user to specify an alternative path
(only relevant for the FileOutput), and allowed to remove the
handlers from the Simulation class.

(Note: the output directory is created when the Ouput object is
constructed, so no handlers can be regsistered by the application
before the 'path exists' problem occurs.