Improved interface of class CrossSection
Various improvements were made to the interface of class CrossSection.
There were some inconsistencies in the names. As an example,
energies() referred to the energie values of the table values from
which the cross section was interpolated (the lookup table), not
to the energies were the cross section is defined. OTOH, the more
explicit raw() was used to retrieve the cross section *values* of
that lookup table. The members are now called rawEnergies() and
rawCrossSection() to make this explicit.
Also various constness fixes have been made. The members above
now return constant references, also getGrid() now returns a constant
pointer to the grid. The data members have been prefixed with m_
to make clear that those are data members, not formal arguments or
local variables. That avoids the somewhat awkward usage of identical
names for arguments and data members in the constructor. An accessor
threshold() was added for the threshold ebergy. All data members
are now private.