Parse: make this a namespace
Parse used to be a structure with static members. There is no good
reason anymore for that, so I changed it to be a namespace with
inline free functions. That is a little bit more straightforward.
Also the specializations of setField are now declared (and implemented)
in the Parse namespace.