use Vector::Index, improve smartGrid construction.
Instead of unit32_t, the Index type that is to be used to address
nodes or cells is now adopted from the (Eigen) Vector type. It is
now defined to be Vector::Index.

For the non-JSON constructor case, the smartGrid pointer is now
set up only if a non-default parameter value is encountered, indicating
that the user specified a smartGrid in the input file. The obsolete
member isSmart of SmartGridParameters could be removed as a result.

The SmartGridParameters constructors now call the new private member
checkConfiguration(), which produces an error if invalid input was
specified.

Various comments have been updated.