Grid: use only one updatedMaxEnergy event.
There used to be two such events, but that is not necessary. The
construction of the electron kinetics class ensures that the
listereners are added (and therefore: executed) in the correct
order. The reason is that the processes/cross sections, which
add listeners in their constructors, are constructed before the
ElectronKinetics class' constructor adds an additional listener.