hessenberg: avoid naked pointer allocations.
Instead of double* c in combination with new and delete, use
C++ facilities. The workspace (pointers c,s,v) has been replaced
with a class HessenbergWorkspace, which implements c,s and v as
a Vector. The code than uses c.data() etc. as storage pointer.
The original hessenberg functions now have the workspace reference
as an additional argument. For compatibility, new functions have
been provided that create the workspace as a local variable, and
pass that on to the overload that accepts the workspace reference.