g_E, g_fieldTemporalGrowth: avoid singularities.
By calculating g_E and g_fieldTemporalGrowth using a good-old loop,
the singularity at u=0 is avoided (this was later repaired y overriding
g[0]=0). The advantage is that we can now enable FPU-exceptions
(see source/main.cpp) to test for the origin of NaN's. The resulting
expressions also allowed the in-place calculation of totalCSI, no
array needs to be set up for that anymore. The result should also be
more efficient, all in all.