Re-implemented removeComments without regex.
The implementation of removeComments was still flawed; as an example
it did not remove trailing whitespace. That was hard to fix, given
the different handling of the $ (end-of-lin? stream?) character,
and the fact that std::regex::multiline is not widely supported.
I therefore re-implemented this using getline and std::string operations,
which is rather straightforward.