Fixed EedfGas::findStatesToUpdate
The recent changes to the state container added a new
'charge' layer in the hierarchy. Now there is a root layer,
followed by a layer that lists any aions, followed by a
layer for the electronic states, etc.
The EedfGas::findStatesToUpdate function was
called on the charge layer instead of the electronic
layer. This is now fixed, the function signature
has also been simplified (no arguments -> use
get_root() and rely on copy/move elision for return).

This should solve #12.