Fixed populations of root and charge levels
Populations of root and (neutral) charge levels were incorrectly
set to one for gasses with fraction==0. This happened in the
constructors, however at the time of loading collisions, gas
fractions have not been loaded. Therefore root and charge populations
are now set in the 'GasBase::propagateFractions' function.
This also removes special casing for electrons.

Should solve #12 for O2
(CO2 still throws an 'energy' related error that was also present
in the previous C++ version)