Allow using system versions of Eigen and nlohmann.
Moved the external codes in different directories that allow
the usage of system versions of these codes (Eigen, nlohmann),
rather than the ones shipped with LoKI-B. Probably we should
get rid of this built-in Eigen completely in the future.

In CMakeLists.txt, have special lines for setting inc_eigen
and inc_nlohmann. Use these to define inc. For now, disable
setting inc_eigen, but still use the built-in nlohmann/json.hpp.

NOTE: the/many/directory/layers/ are needed to allow accidental
picj up of the headers that are shipped with LoKI-B and make clear
what versions these copies have.

This is related to issue #13. We should still decide how to
allow the user to enable/ignore the built-in Eigen and nlohmann,
both in Cmakefiles.txt as well as in configure.ac.