Sanitized the power balance evaluations.
Introduced class PowerTerm that has data members forward and
backward, and a member net() that returns forward+backward.

Use this to implement GasPower and Power. Use PowerTerm also
for ionization and attachment, even if backward==0. That allows
a serious simplification of evaluatePower in the gas and mixture.

The C-style casts from power object pointers to double-pointers
are no longer needed (fortunately, because those could fail
miserable in case of mis-alignment, non-packed arrays.

In output a type was fixed: CAR(gain) appeared twize, the second
must be CAR(loss).