Added notes about the CAR operator.
In the code we see a result proportional to Q_k, instead of to
Q_k^2, as in the papers. The expression in the code is not
backed up by a result written on paper. Also the units must
be checked carefully, as well as the usage of fractions (in the
code) vs. N_k (densities?) in the paper.