Fixed exception safety bug.
Release collision at the moment is is added to the gas, using
eedfGas.addCollision. At that moment, the gas is responsible
for the memory management. Before this patch, when an exception
was thrown in the later actions of createCollision,
collision.release() would not be called. The result is that both
the collision unique_ptr destructor AND the gas would delete the
collision. This happened for the H2 case because the lookup tables
are not correct (energies not in ascending order).

@DAANBOER this fixes the problem that was discussed privately