Provide and use new function maxRelDiff.
Use this function instead of doing, in multiple places in the code,
(eedf - eedfNew).cwiseAbs().array() / eedf.array()).maxCoeff()

At present the function is still implemented like this. When the
define LOKIB_USE_OLD_MAXRELDIFF in LinearALgebra.cpp is commented
out, a new version will be activated that prevents returning NaN
for the case that the two vectors have value 0 for any index.