Enable the new calculation of maxRelDiff(v1,v2).
This avoids a NaN for the case that the two vectors are both 0
for any index. That resulted in convergence not being detected
and excessively long runtimes. The new behaviour is consistent
with how matlab works: that *skips* NaN's in arrays when detecting
a maximum as well. The literal translation into Eigen that was
previously in the code does not.

We now get decent performance also for zero or low E/N values
when a smartGrid is enabled, although the adjustment algorithm
can still be improved.