Re-implemented the state/collision hierarchy.
DO not derived Eedf Gas and Mixture types. Instead, new collision
classes EedfCollisionDataGas and EedfCollisionDataMixture are
available, and a member variables of the latter types is now
managed by the EedfMixture. The collision code is now available
in EedfCollisions.*. There is now a single Gas type (instead of
three), and this is now available in Gas.h.