Advanced Eigen and nlohman-json configuration
find_package(Eigen3) and find_package(nlohmann_json) are called
that attempt to locate system installed Eigen3 and nlohmann_json.
If they are installed in a non-standard location, you can assist
CMake by setting CMAKE_PREFIX_PATH to one (or multiple) folders.
If the package is not found, it is downloaded from gitlab/github.

Alternatively, USE_BUILTIN_EIGEN and USE_BUILTIN_NLOHMANN_JSON
may be set to true/ON/1/etc, which uses the built-in files. An
imported target is set up to keep the interface consistentent.