Simulation: removed configureOutput, m_output.
Instead, the application registers the output task itself with
the m_obtainedEedf event handler of the Simulation class. There
is no reason to special case one particular output. Note that there
is also a print-to-screen handler (that is also registered with
m_obtainedEedf).

I noticed this while documenting the workings of the Simulation
class and the role of the various members. See the doxygen comments
in Simulation.h (and please fix any mistakes I may have made).