Add resilient target_add_test helper function
Use instead of add_test: ctest can not find the tests when they
are relocated after setting CMAKE_RUNTIME_OUTPUT_DIRECTORY. (See
the stackoverflow post mentioned in the CMakeLists).
This problem is solved with generator expressions, and to avoid
repeating the same (awkward) syntax, it is wrapped in a small
function.