Rework getCurrentJobFolder()
In WorkingConditions, introduce getCurrentJobFolder, setCurrentJobFolder
to get a string that is used to identify a case. In Simulation,
after setting the parameters, the result of calling getCurrentJobFolder
on the job manager is passed to WorkingConditions. The Output classes
now use the WorkingConditions' getCurrentJobFolder instead of that
of the job manager. As a result, the Output code has become independent
of the JobManager and this data member/constructor argument has been
removed. The advantage is that it is now possible to do a LoKI-B
calculation without using the job manager while still using the Output
infrastructure. That is important to handle use cases where results for
particular cases are needed on demand, e.g. when LoKI-B is embedded in
another application.

An example that does such direct calculation is presented in file
ideas/loki_in_steps.cpp (at present that does only a single calculation
for E/N=50*Td).

It has been observed by @daanboer that more work could be done in this
area, see the notes in web/bindings.cpp. The whole Output hierarchy
could be made redundant by making more use of 'handleResult' functions.
On the other hand, having stateful output objects has distinct
advantages as well. To be discussed after some more use cases have been
developed.