ElectronKinetics: derive -Boltzmann and -Prescribed
Derive two classes ElectronKineticsBoltzmann and ElectronKineticsPrescribed
from class ElectronKinetics. These will handle the cases of a Boltzmann-based
calculation and the usage of a prescribed EEDF with a shape parameter that
are present in the MATLAB version of the code. (So far, the C++ version
did not support a prescribed EEDF). In this initial changeset:

 - the ElectronKinetics constructor has been made protected,
   so only objects of one of the derived types can be created.
 - the constructors of the derived classes call that of the
   base class, then call the new private member initialize() that
   has been added to both derived classes.
 - The new members initialize() are empty so far, but will contain
   constructor tasks that can be shared by the Setup and JSON constructor
   overloads, just like it is done in the ElectronKinetics base class.
 - Member shapeParameter (and its initialization code) has been moved
   from the ElectronKinetics class to ElectronKineticsPrescribed.
   Only there it will be relevant, so far it was unused. When constructing
   an ElectronKineticsPrescribed object, the presence of 'shapeParameter'
   is now required in the JSON configuration object.

There is quite some code that can be shared between the two classes. As
an example, both will need the fields such as g_CAR to calculate the
various power terms. The difference is that the Boltzmann code also
requires the corresponding matrices (such as CARMatrix) that are terms
of the discretization matrix. Such matrices (and their evaluation in
terms of the g-vectors) will ultimately be moved to the Boltzmann class,
the shared code stays in the ElectronKinetics base class for now.
(In a later stage, the various operators such as the CAR contributions
are better implemented as separate classes that provide interfaces
for updating its g (==g_CAR) and, optionally, also updating a discretization
matrix when one is provided. Then both ElectronKinetics classes can
just use the services of such Operator member in a different way,
without the need for class derivation.)