Event<...>: do not assume arguments are references
Let the user decide if an event template argument is be a reference.
This allows the template arguments to be pointers, and (for example),
passing a nullptr constant. Changed the type definitions of the
ResultEvent in LoKI-B/ElectronKinetics.h accordingly. Also accept
a pointer to the firstAnisotropy instead of a reference, so a nullptr
can be passed to signal that it is not availeble (e.g. in a future
implementation of the 'prescribed' eedf case. Added a test for the
Event template, which also explains its usage.