Implemented the prescribed EEDF class.
Implemented ElectronKineticsPrescribed, reshuffled the code such
that shared code is (still) in the base class. The ElectronKineticsBoltzmann
plugin works as before, the ElectronKineticsPrescribed plugin still
needs work. In particular, it is not clear how the prescribed EEDF
code should interact with other settting (growth) and how it should
handle the field term in the power balance. The MATLAB code calculates
the E/N such that the energy balance is correct, but is is not clear
if/how growth terms and the resulting E/N should affect the calculation
of the swarm parameters. The code will be made to behave just the same
as the MATLAB code, but more documentation of the code and the physics
model is due for this case.