Introduce class FieldOperator
This class manages vector g (representing g_E) and provides
interfaces for setting up the fieldMatrix and evaluating the power
absorbed from the field. The latter member gives the power for the
case without growth (only conservative electron collisions), for
spatial or temporal growth, other or additional terms are present.

An instance fieldOperator of this type is available in the ElectronKinetics
base class, the fieldMatrix has been moved to the derived class
ElectronKineticsBoltzmann, since it is only needed there, not in
the ElectronKineticsPrescribed class.