Move EE members to new class ElectronElectronOperator
New class ElectronElectronOperator can be found in files Operators.*,
which manages the members alphaEE, BAee, A and B that were previously
members of the ElectronKineticsBoltzmann class. It also provides members
evaluatePower, a default constructor that sets alphaEE to zero and
member initialize(const Grid& grid) that sets up the grid-dependent arrays.