Introduce class IonizationOperator
New class IonizationOperator can be found in Operators.* and
contains the code that manages the matrices that appear in the
discretization of the ionization terms. This class is used only
by the ElectronKineticsBoltzmann eedf type; for the prescribed
eedf only the power that is used for (conservative) ionization
events needs to be evaluated, and that is done by calling
evaluatePower on the mixture.collision_data().data_per_gas()
objects.