Simplify control flow when compiling boltzmannMatrix
Do not use separate if-CAR paths. Just add CAR afterwards if it has been
configured. This greatly facilitates the code and should not have a
significant effect on efficiency, since the CAR matrix is tridiagonal and
all these elements exist at the point of addition.

Systematic profiling after the code is feature-complete will reveal which
optimizations are essential.