Formatting, use eeOperator.updateABMatrices.
The formatting of the codes that calculate the boltzmannMatrix has
been changed for better readability. A member updateABMatrices
has been added to class ElectronElectronOperator, it does the task
of updating the matrices and b that are used for the calculation
of A and B in the discretization of the ee-term. The code for doing
this work in ElectronKineticsBoltzmann::solveEEColl has been
replaced by a call to eeOperator.updateABMatrices.