Identified a probably bug in the ee-operator (DB application)
It appears that there is a bug in the implementation of the
ElectronElectronOperator, in the implementation of Rockwood's
detailed balance modification of the matrix a: a temporary
matrix is needed to prevent mixing old and new values. Tests
will follow, and until this has been confirmed, the macro
LOKIB_EE_APPLY_DB_FIX in Operators.cpp is set to 0. When
defined to 1, the fixed algorithm is used. The problem is
explained in the updated notes document.