Fix the detailed balance of ee, add a test.
Add test_ee.cpp. This creates Grid and ElectronElectronOperator
objects from direct parameters, then calculates the flux G_ee
and source *(dG_ee/du)/(N*gamma) for a fixed eedf. In the present
test, a Maxwellian eedf is set up, for which G (and the source)
should be zero. The smallness of the results (compared to a reasonable
reference that is based on the discretisation coefficients) is tested
using test_expr macros (see tests/testUtilities.h).

All tests pass, *because LOKIB_EE_APPLY_DB_FIX has been set to 1*
as part of this patch. Setting it back to 0 results in relative
errors that depend on the grid size (\propto (1/Delta u)^2), whereas
the results are of the order of the machine accuracy with the fix
enabled. The docs have been updated to mention this fact.