Add a demo that shows Maxwellization of an EEDF.
The time-dependent Boltzmann equation is solved for a system of
electrons that only undergo collisions among themselves. The
program shows how the initial (Druyvesteyn) EEDF is Maxwellized
as a result. The program demonstrates how the (conservative)
formulation of the problem results in particle and energy
conservation as time progresses.

The program produces a file 'eedf.dat' with the results at given
moments of time (separated by two empty lines). To show all eedfs
with guplot, you may type (from a gnuplot window):

  set logscale y
  plot for [i=1:*] "eedf.dat" index i title sprintf("%d",i) with lines

(The label is the write-index, not the time.)