ElectronElectronOperator: add member clear()
This sets g_ee and the elements of A and B to zero. This is used
by mixingDirectSolutions to set up the system to solve the growth
matrices, starting without ee collisions. Before the matrices were
cleared directly. Note that the 'internal' matrix BAee is not
cleared. A later patch will ensure that this matrix is updated
if (and only if) the grid changes. At present, this is recalculated
every time solveEEColl is called (also in the MATLAB version),
which is a waste of time and a source of confusion.