Sanitize calls to update_g_ee, updateAB.
In the beginning of the growth functions (solveSpatialGrowthMatrix and
solveTemporalGrowthMatrix, before the convergence loop starts. Member
eeOperator.updateAB was called when EE collisions are enabled. This is
now done at the site of the caller (mixingDirectSolutions). This separates
EE updates from growth model upates and is more clear as a result.

In solveEEColl, follow MATLAB by updating g_ee and AB in the beginning
of the loop, rather than (also) at the end. Also update g_ee *inside*
the loop, since Te may have changed. Before it was calculated only
once, before the loop started. This change has also been made in the
MATLAB code somewhere between versions 1.0.0 and 2.2.0, we are now in
sync again (as far as this issue is concerned). See the related item
in issue #38.

In mixingDirectSolutions, clear() the eeOperator only in case a combined
spatial and EE calculation is going to be done. Setting this also in the
EE-collision-only case does not make sense, since solveEEColl() will
immediately overwrite g_ee and AB.