Call updateABMatrices only when needed.
Before this was called insode solveEEColl. Now member initialize(grid)
is called on the eeOperator in evaluateMatrix instead. This callback
is called when the grid changes (uMax, that is). This member, which
is also called in the constructor of ElectronKineticsBoltzmann,
takes care of updating matrix BAee.

Of course this saves some time. But more importantly, it avoids the
confusion that is caused by the code doing work in places where it
shouldn't.