Make eeOperator a const unique_ptr.
Member eeOperator is now a const unique_ptr<ElectronElectronOperator>.
It is either configured, or set to nullptr in the constructor, depending
on the includeEECollisions input setting. The data member with that
name has been removed, we now test the eeOperator pointer to decide
if ee collisions need to be taken into account. In a few places,
unconditional calls to eeOperator (such as to discretizeTerm(),
which was a no-op if no ee collisions were configured, because all
coefficients were zero) have been guarder with 'if (eeOperator)'.

All in all this saves some time and memory, and makes it much more
difficult to erroneously use functionality that is not available for
a given run.